Run a stack-placed struct argument once, and count a union HFA by its widest member
Three more from the fable-5 review, in its priority order. On arm64 a struct
argument that overflowed the registers was evaluated twice - once by the
scalar loop, which also stored its address where its bytes belonged, and
again by the aggregate loop - so its side effects ran twice. The scalar loop
now skips aggregates. A union of doubles counted as a two-element HFA because
hfaWalk summed members that overlap; it takes the widest now, so the caller
no longer reads sixteen bytes of an eight-byte object. And the constant
folder's unsigned division still truncated through the host's long at two
casts - the class 7188166 removed, finished.

Verified against clang: side effects run once, the union travels in d0 with
the argument after it in d1, arm64 16/16, and a new differential case holds
both down on every target.

+++ src/Parser.cpp
@@ -2221,9 +2221,9 @@ bool Parser::fold(const Expr &e, long long *out, std::size_t pos) const {
                 return true;
             }
             if (b->op() == BinOp::Div)
-                *out = uns ? static_cast<long>(ul / ur) : l / r;
+                *out = uns ? static_cast<long long>(ul / ur) : l / r;
             else
-                *out = uns ? static_cast<long>(ul % ur) : l % r;
+                *out = uns ? static_cast<long long>(ul % ur) : l % r;
             return true;
         case BinOp::Shl:
         case BinOp::Shr:
